zlog: add streaming interface
seqr: specify seq semantics for empty log

reading the sequencer returns the next position to be handed out, that
is, the value returned by the next call to next. an empty log is
identified explicitly, and the first sequence value to be handed out
is position zero. before it would have been 1.

test: fix tail semantics
lib: simplify checktail call

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>