test: fix name clash with struct
This is a bizarre bug I haven't found yet. When aio_state in this patch
is renamed to AioContext there is a segfault in which free is being
applied to an invalid pointer, such as a double free bug. At one point
gdb said it was a bufferlist being freed. The only other name conflict
is in src/libzlog/backend/ceph.cc but that should be hidden as the types
aren't ever in the same scope.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>