kvs: use heap-based after image on txn commit
previously when a txn was committed we (1) wrote the txn after image to
the log by first serializing the heap-baed representation, (2) read and
deserialized the afterimage, and (3) integrated the deserialized
afterimage into the node cache. this commit avoids (2) by waiting until
the txn afterimage is safely written to the log and then folds in the
heap-based after image that is already available.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>