kvs: fix member initialization race
the txn finisher thread needs to be started after the condition variable
it uses is initialized, otherwise there is a race that the thread
sometimes wins and accesses the unitialized cond variable.

from the c++ spec: the non static members are initialized in
declaration order, independent of mem-initializer list order.

fixes: #142

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>