kvs: fix missed condvar wakeup race
txn committed state was set by thread committing the txn, but the txn
handler was examining this state as part of a condition variable held
under a lock that wasn't held when updating the state. so there was a
race where the txn finisher would examined the committed state, and
before blocking on the cond var, the next txn would start and finish,
signal the finisher, all before it blocked so it'd miss the signal.

fixes: #160

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>