kvs: shard the node cache
this is a very basic approach to sharding the LRU cache. there is a lot
more than can be done to make it better, but right out the box, even
with this non-optimised version, we are getting above 100K read ops/sec
with 8 threads. Before every additional thread slowed down overall
throughput.

This patch also eliminates any locks that were previously held during
log I/O which is really useful for mixed workloads. It's a pretty
obvious thing to avoid, but our I/O layer isn't particularly
sophisticated at this point and that locking mistake was overlooked
until we started running benchmarks.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>