dedalus: consider empty objects in seq init
In zlog maxpos returns the maxpos and an empty flag. When all objects
are empty the seq starts with pos 0. Otherwise the sequencer starts with
the value 1 + maxpos across all non-empty objects.

Before in the model we were setting up the contents of the object reply
to work out correctly for empty and non-empty objects without actually
reporting empty state.

Now the model takes empty state into account on the sequencer to arrive
at the initial counter state.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>