dedalus: don't initialize seq until next timestep
Wow. Initializing the sequencer in the same timestep that we receive all
our messages causes assertion errors in molly and c4 to get stuck and
explode its memory usage. This was introduced early on, and seems to
have been non-deterministic, but cropping up now as I fiddled with
various things like larger instances or different pre/post conditions.

What's going on? I suspect that it has to do with the suspicious
recursion between sequencer initiatlization and collecting replies
from objects for max pos.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>