core: unify synchronous / async code paths
This patch removes the aio versions of backend APIs and builds an aio
layer on top of the synchronous versions with a thread pool. This patch
also forces the synchronous versions of the high-level log api to use
the async versions underneath.

This decision is made to simply the code and make testing easier. There
are advantages to using a low-level aio interface, and we can add this
back in at a later time after everything has stabilized.

fixes: #248
fixes: #249

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>