striper: optimistically expand views
this double buffers stripes to avoid a slow path to expand for normal
I/O operations. when a position is mapped we check if there is next
stripe, and if not, we schedule the view to be expanded which then
happens using a worker thread in the background.

a more general form might instead of double buffering stripes, specify
the number of positions that should be pre-allocated.

fixes: #238

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>