log: when proposing new seq wait for winner
this patch forces a thread that encounters a newer epoch during seal to
wait for a newer view. this is super fragile, and needs to change in the
future.

1. out-of-order sealing. currently things work because the first writer
wins and every thread seals in the same order.

2. retrying. if the winning thread crashed before fnishing the
propposal, every other thread will wait forever.

3. optimize the single process case by reducing a thundering herd of
proposers to a single proxy for them all.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>