be/lmdb: persist unique id in db
clients must received unique ids for their secret. in the lmdb
implementation we were not persisting the ids and they always started at
zero even when re-opening a log. the result was that the client observed
observed an old view as if it was one that it created because it would
match the secret.

fixes: #300

Reported-by: Victor Sui <victorsui11@gmail.com>
Signed-off-by: Noah Watkins <noahwatkins@gmail.com>