objectmap: remove const qualifier from members
with const members and the secondary stripe-by-id index using iterators
into the primary stripe-by-pos index we had have a mix of static
initalizer methods for members as well as custom move/copy constructors
for recomputing the index. this was ok, but it was really working
_against_ the original intent of using const members for safety.
instead, const-ness is enforced by requiring all methods to be const,
and reducing the non-const code down to a small part of the constructor.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>