log: do not propose seq when missing
when a sequencer is needed do not propose one on demand. instead, we
propose when the log instance is created. a client can re-purpose if
desired, or exit and restart. normally this is going to be involving
runtime coordination like k8s to determine which process has which role.

when we transition to support shared setup, then it would also be
feasible to instead simply block and wait for a sequencer. there are
numerous variants on these policies that we'll have to explore.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>